# openapi.yaml
openapi: 3.0.3
info:
  title: Sports2 API
  description: |
    REST API for the Sports2 collegiate baseball scouting and team management platform.

    ## Authentication
    All endpoints (except public ones) require:
    - `Authorization: Bearer <JWT>` header
    - `x-csrf-token` header on state-changing requests (POST, PUT, PATCH, DELETE)

    Get a CSRF token first: `GET /api/v1/auth/csrf-token`

    ## Response Format
    Success: `{ "success": true, "data": {...} }`
    Error: `{ "success": false, "error": "..." }`

    ## Scouting System
    See `docs/api/scouting-system-api.md` for the scouting system redesign including Prospects,
    20-80 grading scale, present/future grades, and recruiting pipeline.

    ## Stats & Analytics
    See `docs/api/stats-analytics-api.md` for player season/career stats, videos, leaderboards,
    and news releases (all populated by PrestoSports sync).

    ## Extended Stats & Coach's Dashboard
    See `docs/api/extended-stats-api.md` for split stats, raw Presto data, player/team game logs,
    team aggregate stats, lineup, and the coach's dashboard endpoint.
  version: 1.0.0
  contact:
    name: Sports2 Team
servers:
  - url: http://localhost:5000
    description: Local development (reference only; each deployment has its own endpoint)
security:
  - bearerAuth: []
paths:
...
  /api/v1/tournaments:
    get:
      summary: List tournaments for the team
      tags: [Tournaments]
      parameters:
        - name: season
          in: query
          schema: { type: string }
          description: Filter by season ID
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Paginated list of tournaments with game counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tournament'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /api/v1/tournaments/{id}/games:
    get:
      summary: Get all games in a tournament
      tags: [Tournaments]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
          description: Tournament ID
      responses:
        '200':
          description: Tournament with its games ordered by date
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      tournament:
                        $ref: '#/components/schemas/Tournament'
                      games:
                        type: array
                        items:
                          $ref: '#/components/schemas/Game'
        '404':
          description: Tournament not found
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    csrfToken:
      type: apiKey
      in: header
      name: x-csrf-token
      description: Required for POST, PUT, PATCH, DELETE
  schemas:
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        pages: { type: integer }
...
    Tournament:
      type: object
      properties:
        id: { type: integer }
        team_id: { type: integer }
        name: { type: string, example: 'Snowbird Classic' }
        season: { type: string, nullable: true }
        season_name: { type: string, nullable: true }
        tournament_type: { type: string, enum: [tournament, invitational, scrimmage], nullable: true }
        game_count: { type: integer, description: 'Number of games in this tournament (included in list endpoint)' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Game:
      type: object
      properties:
        id: { type: integer }
        opponent: { type: string }
        game_date: { type: string, format: date-time, nullable: true }
        game_time: { type: string, nullable: true }
        home_away: { type: string, enum: [home, away] }
        team_score: { type: integer, nullable: true }
        opponent_score: { type: integer, nullable: true }
        result: { type: string, enum: [W, L, T], nullable: true }
        location: { type: string, nullable: true }
        opponent_logo_url: { type: string, nullable: true, description: 'Opponent team logo URL from PrestoSports' }
        venue_name: { type: string, nullable: true, description: 'Stadium / field name' }
        event_type: { type: string, nullable: true, description: 'scrimmage | regular' }
        is_conference: { type: boolean, default: false }
        is_neutral: { type: boolean, default: false }
        is_post_season: { type: boolean, default: false }
        tournament_id: { type: integer, nullable: true }
        season: { type: string, nullable: true }
        season_name: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        game_status: { type: string, enum: [scheduled, completed, cancelled, postponed] }
        game_summary: { type: string, nullable: true }
        running_record: { type: string, nullable: true }
        running_conference_record: { type: string, nullable: true }
        source_system: { type: string, enum: [manual, presto] }
...